Adding Media wiki mini (#145)
Summary:
Pull Request resolved: https://github.com/facebookresearch/DCPerf/pull/145

This diff adds a config for media_wiki mini job. I am changing the oss_performance code that will add more options for reducing execution time that needs to be added to the mini job config. So, those options will be added to the config

Also, we need to add --temp-dir to the mini config, but  Ideally, we want to add the --temp-dir, only if the user provide a value for that. So, we need to decide on how to do it.

Reviewed By: excelle08

Differential Revision: D77192531

fbshipit-source-id: 45af99d8ab579106c8fe56672554a51e2785f29d