Add rebatch kernel into Adsim (#181)
Summary:
Pull Request resolved: https://github.com/facebookresearch/DCPerf/pull/181

### Summary

This diff adds a new kernel called "Rebatch" into Adsim. The Rebatch kernel is implemented in the `Rebatch.h` and `Rebatch.cc` files, which are added to the `cea/chips/adsim/cpp2/server/dwarfs` directory. The kernel is also included in the `adsim` BUCK target.

The diff also updates the `CMakeLists.txt` file to include the Rebatch kernel in the build process.

### Changes

* Added `Rebatch.h` and `Rebatch.cc` files to `cea/chips/adsim/cpp2/server/dwarfs` directory
* Added Rebatch kernel to `adsim` BUCK target
* Updated `CMakeLists.txt` file to include Rebatch kernel in build process

### Files Changed

* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/Rebatch.h`
* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/Kernel.cc`
* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/Dwarfs.cc`
* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/Rebatch.cc`
* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/BUCK`
* `fbcode/cea/chips/adsim/BUCK`
* `fbcode/cea/chips/adsim/cpp2/server/dwarfs/CMakeLists.txt`

Reviewed By: excelle08

Differential Revision: D79221719

fbshipit-source-id: 49abdea55eb55ef8d21e33a6f49d05608afdf85d