rebatch: report bidirectional bandwidth
Rebatch currently reports read-only/unidirectional bandwidth for the memcpy operations.

For example, if a memcpy of 1 TiB takes 1 second, Rebatch would report a bandwidth of 1 TiB/s. 
This does not match the behavior of other benchmarks (e.g., STREAM), which would report 2 TiB/s for the same scenario.

Multiplying the memcpy size by 2 solves this discrepancy.